+++ typescript/sdk/src/index.ts
@@ -728,3 +728,4 @@ export {
   getStarknetMailboxContract,
   getStarknetEtherContract,
 } from './utils/starknet.js';
+export { verifyScale } from './utils/decimals.js';
